Rm'd miniconda.
Keep getting this error. Don't want to mess with it.

```
/Users/ramey/miniconda2/bin/python: No module named virtualenvwrapper
virtualenvwrapper.sh: There was a problem running the initialization
hooks.

If Python could not import the module virtualenvwrapper.hook_loader,
check that virtualenvwrapper has been installed for
VIRTUALENVWRAPPER_PYTHON=/Users/ramey/miniconda2/bin/python and that
PATH is
set properly.
```